Update README.md
The objective of this Project is to classify the Iris Flower dataset. Basically 
Iris flower has 3 families namely, Iris Setosa, Iris Virginica and Iris Versicolor, At the end of the project we will be able to recognize the family of the flower by just entering its sepal and petal length and width.